requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
numpy>=1.23.0
yfinance>=0.2.36
sqlalchemy>=2.0.0
plotly>=5.20.0
openpyxl>=3.1.2
tqdm>=4.66.0
ta
python-dotenv
